+++ otlet/api.py
@@ -1,27 +1,29 @@
 """
-otlet.http - otlet http request functions
-CLI tool and wrapper for PyPI JSON Web API
-
-Copyright (c) 2022 Noah Tanner
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
-DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
-OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
-OR OTHER DEALINGS IN THE SOFTWARE.
+otlet.api
+======================
+Functions that call the PyPI Web API.
 """
+#
+# Copyright (c) 2022 Noah Tanner
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+# IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+# DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+# OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
+# OR OTHER DEALINGS IN THE SOFTWARE.
+
 
 import json
 import http.client
@@ -49,7 +51,12 @@ def _attempt_request(url: str) -> Union[http.client.HTTPResponse, PyPIPackageNot
 
 
 def get_full(package: str) -> PackageObject:
-    """Get full response from PyPI API."""
+    """
+    Get full response from PyPI API.
+
+    :param package: Name of package to search for
+    :return: :class:`~otlet.types.PackageObject`
+    """
     res = _attempt_request(f"https://pypi.org/pypi/{package}/json")
     if isinstance(res, PyPIPackageNotFound):
         raise res
@@ -57,7 +64,13 @@ def get_full(package: str) -> PackageObject:
 
 
 def get_release_full(package: str, release: str) -> PackageObject:
-    """Get full response from PyPI API for specific version."""
+    """
+    Get full response from PyPI API for specific version.
+
+    :param package: Name of package to search for
+    :param release: Release version of package
+    :return: :class:`~otlet.types.PackageObject`
+    """
     res = _attempt_request(f"https://pypi.org/pypi/{package}/{release}/json")
     if isinstance(res, PyPIPackageNotFound):
         res = _attempt_request(
@@ -73,7 +86,15 @@ def get_release_full(package: str, release: str) -> PackageObject:
 
 
 def get_info(package: str) -> PackageInfoObject:
-    """Get package info from PyPI API."""
+    """
+    Get package info from PyPI API.
+
+    :param package: Name of package to search for
+    :return: :class:`~otlet.types.PackageInfoObject`
+
+    .. deprecated:: 0.3.0
+        This function is deprecated and will be removed in version 1.0.0
+    """
     deprecated(
         "get_info and get_release_info are deprecated methods and will be removed in version 1.0.0"
     )
@@ -82,7 +103,16 @@ def get_info(package: str) -> PackageInfoObject:
 
 
 def get_release_info(package: str, release: str) -> PackageInfoObject:
-    """Get package info from PyPI API for specific version."""
+    """
+    Get package info from PyPI API for specific version.
+
+    :param package: Name of package to search for
+    :param release: Release version of package
+    :return: :class:`~otlet.types.PackageInfoObject`
+
+    .. deprecated:: 0.3.0
+        This function is deprecated and will be removed in version 1.0.0
+    """
     pkg = get_release_full(package, release)
     deprecated(
         "get_info and get_release_info are deprecated methods and will be removed in version 1.0.0"